flate2 dependency now optional
Renamed dependency from 'flate2' to 'flate' and made it optional.
Created new module 'archive' which contains all the previously existing
api concerning the compressing and extracting data from archives.
Further more, the test cases for arhived data were removed and the data
was extracted manualy as it should not concern this library of whether
flate2 extracts data correctly.

Motivation: To make the built-in archive functionality optional and
opt-in. This is for the cases when people want to use more advanced
versions of such dependency and just would prefer to skip this
compilation step.